EW-755 separating classes
